# swagger_poolhub.yaml
openapi: 3.0.3
info:
  title: PoolHub Entertainment Center Management API
  description: |-
    Đặc tả hệ thống RESTful API cho **PoolHub - Hệ thống quản lý trung tâm giải trí bi-a và dịch vụ** phục vụ môn học **PRN232 - Tuần 2**.
    
    Tài liệu này định nghĩa 100% tất cả các endpoint nghiệp vụ thực tế khớp với file thiết kế **Analysis PoolHub.xlsx**, hỗ trợ phân quyền truy cập chặt chẽ (Guest, Staff, Manager, Admin) và mô hình phản hồi JSON chuẩn hóa.
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication & Settings
      summary: Đăng ký nhân viên (Admin Only)
      description: Tạo tài khoản nhân viên mới (Staff hoặc Manager) dưới quyền quản trị viên Admin.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Đăng ký thành công tài khoản
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Dữ liệu đầu vào không hợp lệ hoặc bị trùng username/email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Chưa xác thực token JWT
        '403':
          description: Không đủ quyền Admin
  /auth/login:
    post:
      tags:
        - Authentication & Settings
      summary: Đăng nhập hệ thống (Public)
      description: Xác thực thông tin đăng nhập và trả về JWT Access Token để sử dụng các API bảo mật.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Sai email hoặc mật khẩu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags:
        - Authentication & Settings
      summary: Đăng xuất hệ thống
      description: Vô hiệu hóa access token và xóa JWT token lưu ở cookie.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Đăng xuất thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Đăng xuất thành công.
...
  /auth/me:
    get:
      tags:
        - Authentication & Settings
      summary: Lấy profile người dùng hiện tại
      description: Trả về thông tin chi tiết người dùng đang đăng nhập dựa trên JWT token được gửi kèm.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lấy profile thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Token không hợp lệ hoặc hết hạn
  /auth/change-password:
    post:
      tags:
        - Authentication & Settings
      summary: Đổi mật khẩu tài khoản
      description: Cho phép người dùng đang đăng nhập thay đổi mật khẩu của mình.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Đổi mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Đổi mật khẩu thành công.
        '400':
          description: Mật khẩu cũ không chính xác hoặc mật khẩu mới quá ngắn
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - fullName
        - password
        - role
      properties:
        username:
          type: string
          example: staff_nguyena
        email:
          type: string
          example: nguyena@poolhub.vn
        fullName:
          type: string
          example: Nguyen Van A
        phone:
          type: string
          example: "0987654321"
        password:
          type: string
          example: Password123@
        role:
          type: string
          enum: [Staff, Manager, Admin]
          example: Staff
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Đăng ký tài khoản nhân viên thành công.
        data:
          type: object
          properties:
            userId:
              type: string
              example: usr_9f8e7d6c
            username:
              type: string
              example: staff_nguyena
            email:
              type: string
              example: nguyena@poolhub.vn
            role:
              type: string
              example: Staff
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: nguyena@poolhub.vn
        password:
          type: string
          example: StrongPassword123@
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Đăng nhập thành công.
        data:
          type: object
          properties:
            accessToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            expiresIn:
              type: integer
              example: 28800
            user:
              type: object
              properties:
                userId:
                  type: string
                  example: usr_9f8e7d6c
                fullName:
                  type: string
                  example: Nguyen Van A
                email:
                  type: string
                  example: nguyena@poolhub.vn
                role:
                  type: string
                  example: Staff
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            userId:
              type: string
              example: usr_9f8e7d6c
            fullName:
              type: string
              example: Nguyen Van A
            email:
              type: string
              example: nguyena@poolhub.vn
            role:
              type: string
              example: Staff
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          example: OldPassword123@
        newPassword:
          type: string
          example: NewPassword123@
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validate dữ liệu thất bại.
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Định dạng email không chính xác.